requirements.txt
aiogram==3.12.0
aiosqlite==0.20.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytz==2023.3
Pillow==10.1.0